Extract printing plan check
